move older sidekick.js and comment it out
